pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "address_audit"
version = "0.1.0"
description = "地址稽核"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "tanxiaobing", email = "hualing314159@126.com" }]

dependencies = [
    "duckdb>=0.10.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "pyyaml>=6.0.0",
    "sqlglot>=23.0.0",
    "pydantic>=2.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "prefect>=2.14.0",
    "dbt-core>=1.7.0",
    "dbt-duckdb>=1.7.0",
    "httpx>=0.27.0",
    "openpyxl>=3.1.5",
    "python-dotenv==1.0.1",
]

# 可选依赖：语义层（按需启用）
[project.optional-dependencies]
metricflow = ["metricflow>=0.202.0"]

[tool.hatch.build.targets.wheel]
packages = ["address_audit"]


